Make setup_docker.sh idempotent

+++ extras/setup_docker.sh
@@ -1,22 +1,31 @@
 #!/usr/bin/env bash
 
 
-# Uninstall old versions
-sudo apt remove docker docker-engine docker.io containerd runc
+command -v docker >/dev/null 2>&1 && {
+	echo "Uninstalling old version of docker"
+	sudo apt remove docker --yes
+	sudo apt remove docker-engine --yes
+	sudo apt remove docker.io --yes
+	sudo apt remove containerd --yes
+	sudo apt remove runc --yes
+}
 
 
 # Set up the repository
 sudo apt install ca-certificates curl gnupg lsb-release --yes
 
-sudo mkdir -p /etc/apt/keyrings
+if [ ! -f /etc/apt/keyrings/docker.gpg ]; then
+	sudo mkdir -p /etc/apt/keyrings
 
-curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+	curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
 
-sudo chmod a+r /etc/apt/keyrings/docker.gpg
+	sudo chmod a+r /etc/apt/keyrings/docker.gpg
+fi
 
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
-  $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+if [ ! -f /etc/apt/sources.list.d/docker.list ]; then
+	echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian bullseye stable" \
+	| sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+fi
 
 
 # Install Docker Engine
